changes on theory part of auto-boxing and unboxing section

+++ Auto-Boxing and UnBoxing/Example/src/Main.java
@@ -6,9 +6,9 @@ public class Main
 		ArrayList<Integer> intList = new ArrayList<Integer>();
 		
 		for( int loop = 1; loop <= 10; loop++ )
-			intList.add( Integer.valueOf( loop ) ); // Auto-Boxing
+			intList.add( loop ); // Auto-Boxing done, actually we need to enter 'Integer.valueOf( loop )'
 		
 		for( int loopVar = 0; loopVar < intList.size(); loopVar++ ) 
-			System.out.println( (loopVar + 1) + " is " + intList.get( loopVar ).intValue() ); // Un-boxing
+			System.out.println( (loopVar + 1) + " is " + intList.get( loopVar ) ); // Un-boxing done, actually we need to enter 'intList.get( loopVar ).intValue()'
 	}
 }
